refactor call generation code for CommandData and CommandGene
